refactor(ui): create ui/ sub-package — 8 files (slice 5/8, #170)
