+++ src/voicecli/cli_dictate.py
@@ -161,13 +161,13 @@ def dictate(
 
     if listen:
         from voicecli.config import load_stt_config
-        from voicecli.stt_client import hotkey_loop
+        from voicecli.ui.stt_client import hotkey_loop
 
         stt_cfg = load_stt_config()
         hotkey_loop(stt_cfg["hotkey"], paste=paste)
         return
 
-    from voicecli.stt_client import auto_paste, notify, send_toggle
+    from voicecli.ui.stt_client import auto_paste, notify, send_toggle
 
     resp = send_toggle(mode=mode)
 
@@ -200,7 +200,7 @@ def dictate(
 @dictate_app.command("status")
 def dictate_status() -> None:
     """Show current STT daemon state."""
-    from voicecli.stt_client import send_status
+    from voicecli.ui.stt_client import send_status
 
     resp = send_status()
     if resp.get("status") == "error":
@@ -227,7 +227,7 @@ def dictate_test_overlay() -> None:
 @dictate_app.command("next-mode")
 def dictate_next_mode() -> None:
     """Cycle to the next STT mode (becomes the new default)."""
-    from voicecli.stt_client import notify, send_next_mode
+    from voicecli.ui.stt_client import notify, send_next_mode
 
     resp = send_next_mode()
     if resp.get("status") == "error":
@@ -242,7 +242,7 @@ def dictate_next_mode() -> None:
 @dictate_app.command("cancel")
 def dictate_cancel() -> None:
     """Cancel the current STT recording without transcribing."""
-    from voicecli.stt_client import send_cancel
+    from voicecli.ui.stt_client import send_cancel
 
     resp = send_cancel()
     if resp.get("status") == "error":
@@ -287,7 +287,7 @@ def dictate_history(
     """Show the last 20 dictation history entries."""
     import json as _json
 
-    from voicecli.clipboard import write_clipboard
+    from voicecli.ui.clipboard import write_clipboard
     from voicecli.history import HISTORY_PATH
 
     if not HISTORY_PATH.exists():
@@ -390,18 +390,18 @@ def dictate_nats(
     """
     import asyncio
 
-    from voicecli.clipboard import write_clipboard
+    from voicecli.ui.clipboard import write_clipboard
     from voicecli.config import (
         apply_nats_env_from_config,
         load_config,
         load_vocab,
         vocab_to_prompt,
     )
-    from voicecli.nats_recorder import is_recording, start_recording, stop_recording
+    from voicecli.ui.nats_recorder import is_recording, start_recording, stop_recording
     from voicecli.nats_stt_client import transcribe_via_nats
-    from voicecli.stt_client import notify
+    from voicecli.ui.stt_client import notify
     from voicecli.stt_modes import get_mode
-    from voicecli.ui_sounds import play_ui_sound
+    from voicecli.ui.sounds import play_ui_sound
 
     apply_nats_env_from_config()
 
@@ -471,7 +471,7 @@ def dictate_nats(
         notify(f"{lang_tag}{preview}", timeout=3000)
 
         if paste:
-            from voicecli.clipboard import auto_paste
+            from voicecli.ui.clipboard import auto_paste
 
             auto_paste()
     else:
