+++ src/voicecli/runtime/stt_daemon.py
@@ -22,13 +22,13 @@ from pathlib import Path
 
 from roxabi_nats import sanitize_for_wire
 
-from voicecli.clipboard import auto_paste, write_clipboard
+from voicecli.ui.clipboard import auto_paste, write_clipboard
 from voicecli.config import load_stt_config
 from voicecli.runtime.wire_protocol import recv_json
 from voicecli.runtime.wire_protocol import send_json as _send_json
 from voicecli.history import append_history, wav_duration_s
 from voicecli.paths import STT_SOCKET_PATH as SOCKET_PATH
-from voicecli.ui_sounds import play_ui_sound
+from voicecli.ui.sounds import play_ui_sound
 
 MAX_MSG = 65536
