+++ src/voicecli/ui/__init__.py
@@ -1,0 +1,1 @@
+# ui sub-package — overlay, sounds, clipboard, STT client, NATS recorder, Telegram
